requirements.txt
streamlit
PyPDF2
google-generativeai
python-dotenv
scikit-learn
pandas
numpy
nltk
